+++ R/parse_python.R
@@ -198,6 +198,15 @@ expr_has_vars <- function(x) {
 #'
 #' @param x caracas_symbol
 #' @param first_doit Try `doit()` first
+#' 
+#' @examples 
+#' if (has_sympy()) {
+#'   v <- vector_sym(2)
+#'   x <- as_expr(v)
+#'   x
+#'   y <- as.expression(v)
+#'   y
+#' }
 #'
 #' @concept caracas_symbol
 #'
@@ -253,3 +262,24 @@ as_expr.caracas_symbol <- function(x, first_doit = TRUE) {
   }
 }
 
+
+
+#' @export
+as_expr.caracas_solve_sys_sol <- function(x, first_doit = TRUE) {
+  out <- lapply(x, lapply, as_expr, first_doit)
+  return(out)
+}
+
+
+
+#' @inherit as_expr
+#' @export
+as.expression.caracas_symbol <- function(x, ...) {
+  as_expr(x, ...)
+}
+
+#' @inherit as_expr
+#' @export
+as.expression.caracas_solve_sys_sol <- function(x, ...) {
+  as_expr(x, ...)
+}